LPS-77317 user-associated-data-api - changes signature on 'count' method to return an int instead of a long

+++ modules/apps/foundation/user-associated-data/user-associated-data-api/src/main/java/com/liferay/user/associated/data/aggregator/BaseUADEntityAggregator.java
@@ -25,7 +25,7 @@ import java.util.List;
 public abstract class BaseUADEntityAggregator implements UADEntityAggregator {
 
 	@Override
-	public long count(long userId) {
+	public int count(long userId) {
 		List<UADEntity> userIdUADEntities = getUADEntities(userId);
 
 		return userIdUADEntities.size();
